+++ website/src/components/index.js
@@ -1,4 +1,2 @@
-export {default as makeExample} from './example/make-example';
 
-export {default as Home} from './home';
-export {default as InfoPanel} from './info-panel';
+export {default as HomeExplorer} from './explorer';
